+++ src/i18n/langs.ts
@@ -1,0 +1,7 @@
+import * as en from './en.json';
+import * as zh_cn from './zh-cn.json';
+
+export const LANGUAGES = {
+  en,
+  zh_cn
+};