+++ src/settings.ts
@@ -1,6 +1,8 @@
 import { App, PluginSettingTab, Setting } from 'obsidian';
 import WordpressPlugin from './main';
-import { PostStatus } from './wp-api';
+import { CommentStatus, PostStatus } from './wp-api';
+import { LanguageWithAuto, TranslateKey } from './i18n';
+
 
 export const enum ApiType {
   XML_RPC = 'xml-rpc',
@@ -8,6 +10,11 @@ export const enum ApiType {
 }
 
 export interface WordpressPluginSettings {
+
+  /**
+   * Plugin language.
+   */
+  lang: LanguageWithAuto;
 
   /**
    * API type.
@@ -20,9 +27,14 @@ export interface WordpressPluginSettings {
   endpoint: string;
 
   /**
-   * WordPress user name.
+   * XML-RPC path.
    */
-  userName?: string;
+  xmlRpcPath: string;
+
+  /**
+   * WordPress username.
+   */
+  username?: string;
 
   /**
    * WordPress password.
@@ -30,9 +42,9 @@ export interface WordpressPluginSettings {
   password?: string;
 
   /**
-   * Save user name to local data.
+   * Save username to local data.
    */
-  saveUserName: boolean;
+  saveUsername: boolean;
 
   /**
    * Save user password to local data.
@@ -48,15 +60,23 @@ export interface WordpressPluginSettings {
    * Default post status.
    */
   defaultPostStatus: PostStatus;
+
+  /**
+   * Default comment status.
+   */
+  defaultCommentStatus: CommentStatus;
 }
 
 export const DEFAULT_SETTINGS: WordpressPluginSettings = {
+  lang: 'auto',
   apiType: ApiType.XML_RPC,
   endpoint: '',
-  saveUserName: false,
+  xmlRpcPath: '/xmlrpc.php',
+  saveUsername: false,
   savePassword: false,
   showRibbonIcon: false,
-  defaultPostStatus: PostStatus.Draft
+  defaultPostStatus: PostStatus.Draft,
+  defaultCommentStatus: CommentStatus.Open
 }
 
 export class WordpressSettingTab extends PluginSettingTab {
@@ -69,29 +89,33 @@ export class WordpressSettingTab extends PluginSettingTab {
 	}
 
 	display(): void {
+    const t = (key: TranslateKey, vars?: Record<string, string>): string => {
+      return this.plugin.i18n.t(key, vars);
+    };
+
 		const { containerEl } = this;
 
 		containerEl.empty();
 
-    containerEl.createEl('h2', { text: 'Settings for WordPress plugin' });
+    containerEl.createEl('h1', { text: t('settings_title') });
 
 		new Setting(containerEl)
-			.setName('WordPress URL')
-			.setDesc('Full path of installed WordPress, for example, https://example.com/wordpress')
+			.setName(t('settings_url'))
+			.setDesc(t('settings_urlDesc'))
 			.addText(text => text
-				.setPlaceholder('https://example.com/wordpress')
+				.setPlaceholder(t('settings_urlPlaceholder'))
 				.setValue(this.plugin.settings.endpoint)
 				.onChange(async (value) => {
           this.plugin.settings.endpoint = value;
           await this.plugin.saveSettings();
         }));
     new Setting(containerEl)
-      .setName('API Type')
-      .setDesc('Select which API you want to use.')
+      .setName(t('settings_apiType'))
+      .setDesc(t('settings_apiTypeDesc'))
       .addDropdown((dropdown) => {
         dropdown
-          .addOption(ApiType.XML_RPC, 'XML-RPC')
-          .addOption(ApiType.RestAPI_miniOrange, 'REST API Authentication by miniOrange')
+          .addOption(ApiType.XML_RPC, t('settings_apiTypeXmlRpc'))
+          .addOption(ApiType.RestAPI_miniOrange, t('settings_apiTypeRestMiniOrange'))
           .setValue(this.plugin.settings.apiType)
           .onChange(async (value: ApiType) => {
             this.plugin.settings.apiType = value;
@@ -99,36 +123,61 @@ export class WordpressSettingTab extends PluginSettingTab {
             this.display();
           });
       });
+    if (this.plugin.settings.apiType === ApiType.XML_RPC) {
+      new Setting(containerEl)
+        .setName(t('settings_xmlRpcPath'))
+        .setDesc(t('settings_xmlRpcPathDesc'))
+        .addText(text => text
+          .setPlaceholder('/xmlrpc.php')
+          .setValue(this.plugin.settings.xmlRpcPath)
+          .onChange(async (value) => {
+            this.plugin.settings.xmlRpcPath = value;
+            await this.plugin.saveSettings();
+          }));
+    }
     new Setting(containerEl)
-      .setName('Show icon in sidebar')
-      .setDesc(`If enabled, a button which opens publish panel will be added to the Obsidian sidebar.
-Changes only take effect on reload.`)
+      .setName(t('settings_showRibbonIcon'))
+      .setDesc(t('settings_showRibbonIconDesc'))
       .addToggle((toggle) =>
         toggle
           .setValue(this.plugin.settings.showRibbonIcon)
           .onChange(async (value) => {
             this.plugin.settings.showRibbonIcon = value;
             await this.plugin.saveSettings();
-            this.display();
 
             this.plugin.updateRibbonIcon();
           }),
       );
 
     new Setting(containerEl)
-      .setName('Default Post Status')
-      .setDesc('Post status which will be published to WordPress.')
+      .setName(t('settings_defaultPostStatus'))
+      .setDesc(t('settings_defaultPostStatusDesc'))
       .addDropdown((dropdown) => {
         dropdown
-          .addOption(PostStatus.Draft, 'draft')
-          .addOption(PostStatus.Publish, 'publish')
+          .addOption(PostStatus.Draft, t('settings_defaultPostStatusDraft'))
+          .addOption(PostStatus.Publish, t('settings_defaultPostStatusPublish'))
           // .addOption(PostStatus.Future, 'future')
           .setValue(this.plugin.settings.defaultPostStatus)
           .onChange(async (value: PostStatus) => {
             this.plugin.settings.defaultPostStatus = value;
             await this.plugin.saveSettings();
-            this.display();
+          });
+      });
+
+    new Setting(containerEl)
+      .setName(t('settings_defaultPostComment'))
+      .setDesc(t('settings_defaultPostCommentDesc'))
+      .addDropdown((dropdown) => {
+        dropdown
+          .addOption(CommentStatus.Open, t('settings_defaultPostCommentOpen'))
+          .addOption(CommentStatus.Closed, t('settings_defaultPostCommentClosed'))
+          // .addOption(PostStatus.Future, 'future')
+          .setValue(this.plugin.settings.defaultCommentStatus)
+          .onChange(async (value: CommentStatus) => {
+            this.plugin.settings.defaultCommentStatus = value;
+            await this.plugin.saveSettings();
           });
       });
 	}
+
 }