+++ src/wp-publish-modal.ts
@@ -1,8 +1,9 @@
 import { App, Modal, Setting } from 'obsidian';
 import WordpressPlugin from './main';
 import { WordPressPostParams } from './wp-client';
-import { PostStatus, Term } from './wp-api';
+import { CommentStatus, PostStatus, Term } from './wp-api';
 import { toNumber } from 'lodash-es';
+import { TranslateKey } from './i18n';
 
 /**
  * WordPress publish modal.
@@ -19,44 +20,60 @@ export class WpPublishModal extends Modal {
   }
 
   onOpen() {
+    const t = (key: TranslateKey, vars?: Record<string, string>): string => {
+      return this.plugin.i18n.t(key, vars);
+    };
+
     const params: WordPressPostParams = {
       status: this.plugin.settings.defaultPostStatus,
+      commentStatus: this.plugin.settings.defaultCommentStatus,
       categories: [ 1 ]
     };
 
     const { contentEl } = this;
 
-    contentEl.createEl('h1', { text: 'Publish to WordPress' });
+    contentEl.createEl('h1', { text: t('publishModal_title') });
 
     new Setting(contentEl)
-      .setName('Post Status')
+      .setName(t('publishModal_postStatus'))
       .addDropdown((dropdown) => {
         dropdown
-          .addOption(PostStatus.Draft, 'draft')
-          .addOption(PostStatus.Publish, 'publish')
+          .addOption(PostStatus.Draft, t('publishModal_postStatusDraft'))
+          .addOption(PostStatus.Publish, t('publishModal_postStatusPublish'))
           // .addOption(PostStatus.Future, 'future')
           .setValue(this.plugin.settings.defaultPostStatus)
-          .onChange(async (value: PostStatus) => {
+          .onChange((value: PostStatus) => {
             params.status = value;
+          });
+      });
+    new Setting(contentEl)
+      .setName(t('publishModal_commentStatus'))
+      .addDropdown((dropdown) => {
+        dropdown
+          .addOption(CommentStatus.Open, t('publishModal_commentStatusOpen'))
+          .addOption(CommentStatus.Closed, t('publishModal_commentStatusClosed'))
+          .setValue(this.plugin.settings.defaultCommentStatus)
+          .onChange((value: CommentStatus) => {
+            params.commentStatus = value;
           });
       });
     if (this.categories.length > 0) {
       new Setting(contentEl)
-        .setName('Category')
+        .setName(t('publishModal_category'))
         .addDropdown((dropdown) => {
           this.categories.forEach(it => {
             dropdown.addOption(it.id, it.name);
           });
           dropdown
             .setValue("1")
-            .onChange(async (value: string) => {
+            .onChange((value) => {
               params.categories = [ toNumber(value) ];
             });
         });
     }
     new Setting(contentEl)
       .addButton(button => button
-        .setButtonText('Publish')
+        .setButtonText(t('publishModal_publishButtonText'))
         .setCta()
         .onClick(() => {
           this.onSubmit(params, this);