+++ src/wp-rest-client.ts
@@ -1,16 +1,18 @@
-import { App, requestUrl } from 'obsidian';
-import { WordPressClientResult, WordPressClientReturnCode, WordPressPostParams } from './wp-client';
+import { App } from 'obsidian';
+import {
+  WordPressAuthParams,
+  WordPressClientResult,
+  WordPressClientReturnCode,
+  WordPressPostParams
+} from './wp-client';
 import { AbstractWordPressClient } from './abstract-wp-client';
 import WordpressPlugin from './main';
 import { Term } from './wp-api';
-
-interface RestOptions {
-  url: URL;
-}
+import { RestClient } from './rest-client';
 
 export class WpRestClient extends AbstractWordPressClient {
 
-  private readonly options: RestOptions;
+  private readonly client: RestClient;
 
   constructor(
     readonly app: App,
@@ -18,21 +20,19 @@ export class WpRestClient extends AbstractWordPressClient {
     private readonly context: WpRestClientContext
   ) {
     super(app, plugin);
-    this.options = {
+    this.client = new RestClient({
       url: new URL(plugin.settings.endpoint)
-    };
+    });
   }
 
-  publish(title: string, content: string, postParams: WordPressPostParams, wp: {
-    userName: string,
-    password: string
-  }): Promise<WordPressClientResult> {
-    return this.httpPost(
+  publish(title: string, content: string, postParams: WordPressPostParams, wp: WordPressAuthParams): Promise<WordPressClientResult> {
+    return this.client.httpPost(
       'wp-json/wp/v2/posts',
       {
         title,
         content,
         status: postParams.status,
+        comment_status: postParams.commentStatus,
         categories: postParams.categories
       },
       {
@@ -65,8 +65,8 @@ export class WpRestClient extends AbstractWordPressClient {
       });
   }
 
-  getCategories(wp: { userName: string; password: string }): Promise<Term[]> {
-    return this.httpGet(
+  getCategories(wp: WordPressAuthParams): Promise<Term[]> {
+    return this.client.httpGet(
       'wp-json/wp/v2/categories',
       {
         headers: this.context.getHeaders(wp)
@@ -74,56 +74,23 @@ export class WpRestClient extends AbstractWordPressClient {
       .then(data => data as Term[] ?? []);
   }
 
-  protected httpGet(
-    path: string,
-    options?: {
-      headers: Record<string, string>
-    }
-  ): Promise<unknown> {
-    const opts = {
-      headers: {},
-      ...options
-    }
-    console.log('REST GET', `${this.options.url.toString()}${path}`, opts);
-    return requestUrl({
-      url: `${this.options.url.toString()}${path}`,
-      method: 'GET',
-      headers: {
-        'Content-Type': 'application/json',
-        'User-Agent': 'obsidian.md',
-        ...opts.headers
-      }
-    })
-      .then(response => {
-        console.log('GET response', response);
-        return response.json;
-      });
-  }
-
-  protected httpPost(
-    path: string,
-    body: unknown,
-    options?: {
-      headers: Record<string, string>
-    }): Promise<unknown> {
-    const opts = {
-      headers: {},
-      ...options
-    }
-    console.log('REST POST', `${this.options.url.toString()}${path}`, opts);
-    return requestUrl({
-      url: `${this.options.url.toString()}${path}`,
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'User-Agent': 'obsidian.md',
-        ...opts.headers
-      },
-      body: JSON.stringify(body)
-    })
-      .then(response => {
-        console.log('POST response', response);
-        return response.json;
+  checkUser(certificate: WordPressAuthParams): Promise<WordPressClientResult> {
+    return this.client.httpGet(
+      `wp-json/wp/v2/users/?username=${certificate.username}`,
+      {
+        headers: this.context.getHeaders(certificate)
+      })
+      .then(data => {
+        return {
+          code: WordPressClientReturnCode.OK,
+          data: data
+        };
+      })
+      .catch(error => {
+        return {
+          code: WordPressClientReturnCode.Error,
+          data: this.plugin.i18n.t('error_invalidUser')
+        };
       });
   }
 
@@ -132,15 +99,15 @@ export class WpRestClient extends AbstractWordPressClient {
 interface WpRestClientContext {
   name: string;
 
-  getHeaders(wp: { userName: string, password: string }): Record<string, string>;
+  getHeaders(wp: WordPressAuthParams): Record<string, string>;
 }
 
 export class WpRestClientMiniOrangeContext implements WpRestClientContext {
   name: 'WpRestClientMiniOrangeContext';
 
-  getHeaders(wp: { userName: string, password: string }): Record<string, string> {
+  getHeaders(wp: WordPressAuthParams): Record<string, string> {
     return {
-      'Authorization': `Basic ${Buffer.from(`${wp.userName}:${wp.password}`).toString('base64')}`
+      'Authorization': `Basic ${Buffer.from(`${wp.username}:${wp.password}`).toString('base64')}`
     }
   }
 }