+++ src/wp-xml-rpc-client.ts
@@ -1,9 +1,23 @@
-import { App } from 'obsidian';
+import { App, Notice } from 'obsidian';
 import WordpressPlugin from './main';
-import { WordPressClientResult, WordPressClientReturnCode, WordPressPostParams } from './wp-client';
+import {
+  WordPressAuthParams,
+  WordPressClientResult,
+  WordPressClientReturnCode,
+  WordPressPostParams
+} from './wp-client';
 import { XmlRpcClient } from './xmlrpc-client';
 import { AbstractWordPressClient } from './abstract-wp-client';
 import { Term } from './wp-api';
+
+interface FaultResponse {
+  faultCode: string;
+  faultString: string;
+}
+
+function isFaultResponse(response: unknown): response is FaultResponse {
+  return (response as FaultResponse).faultCode !== undefined;
+}
 
 export class WpXmlRpcClient extends AbstractWordPressClient {
 
@@ -15,21 +29,20 @@ export class WpXmlRpcClient extends AbstractWordPressClient {
   ) {
     super(app, plugin);
     this.client = new XmlRpcClient({
-      url: new URL(plugin.settings.endpoint)
+      url: new URL(plugin.settings.endpoint),
+      xmlRpcPath: plugin.settings.xmlRpcPath
     });
   }
 
-  publish(title: string, content: string, postParams: WordPressPostParams, wp: {
-    userName: string,
-    password: string
-  }): Promise<WordPressClientResult> {
+  publish(title: string, content: string, postParams: WordPressPostParams, wp: WordPressAuthParams): Promise<WordPressClientResult> {
     return this.client.methodCall('wp.newPost', [
       0,
-      wp.userName,
+      wp.username,
       wp.password,
       {
         post_type: 'post',
         post_status: postParams.status,
+        comment_status: postParams.commentStatus,
         post_title: title,
         post_content: content,
         terms: {
@@ -37,9 +50,8 @@ export class WpXmlRpcClient extends AbstractWordPressClient {
         }
       }
     ])
-      .then((response: any) => { // eslint-disable-line
-        if (response.faultCode && response.faultString) {
-          // it means error happens
+      .then(response => {
+        if (isFaultResponse(response)) {
           return {
             code: WordPressClientReturnCode.Error,
             data: {
@@ -47,27 +59,55 @@ export class WpXmlRpcClient extends AbstractWordPressClient {
               message: response.faultString
             }
           };
-        } else {
-          return {
-            code: WordPressClientReturnCode.OK,
-            data: response
-          }
         }
+        return {
+          code: WordPressClientReturnCode.OK,
+          data: response
+        };
       });
   }
 
-  getCategories(wp: { userName: string; password: string }): Promise<Term[]> {
+  getCategories(wp: WordPressAuthParams): Promise<Term[]> {
     return this.client.methodCall('wp.getTerms', [
       0,
-      wp.userName,
+      wp.username,
       wp.password,
       'category'
     ])
+      .then(response => {
+        if (isFaultResponse(response)) {
+          const fault = `${response.faultCode}: ${response.faultString}`;
+          new Notice(fault);
+          throw new Error(fault);
+        }
+        return response;
+      })
       .then((data: unknown[]) => {
         return data.map((it: any) => ({
           ...it,
           id: it.term_id
         })) ?? [];
+      });
+  }
+
+  checkUser(certificate: WordPressAuthParams): Promise<WordPressClientResult> {
+    return this.client.methodCall('wp.getProfile', [
+      0,
+      certificate.username,
+      certificate.password
+    ])
+      .then(response => {
+        if (isFaultResponse(response)) {
+          return {
+            code: WordPressClientReturnCode.Error,
+            data: `${response.faultCode}: ${response.faultString}`
+          };
+        } else {
+          return {
+            code: WordPressClientReturnCode.OK,
+            data: response
+          };
+        }
       });
   }